+++ packages/core/src/createMachine.ts
@@ -9,14 +9,60 @@ import {
   AnyEventObject,
   Cast,
   InternalMachineImplementations,
+  IsNever,
   MachineConfig,
   MachineContext,
   NonReducibleUnknown,
   ParameterizedObject,
   Prop,
   ProvidedActor,
+  StateValue,
   ToChildren
 } from './types.ts';
+
+type TestValue =
+  | string
+  | {
+      [k: string]: TestValue | undefined;
+    };
+
+type _GroupTestValues<TTestValue extends string | TestValue> =
+  TTestValue extends string
+    ? TTestValue extends `${string}.${string}`
+      ? [never, never]
+      : [TTestValue, never]
+    : [never, TTestValue];
+type GroupTestValues<TTestValue extends string | TestValue> = {
+  leafCandidates: _GroupTestValues<TTestValue>[0];
+  nonLeaf: _GroupTestValues<TTestValue>[1];
+};
+
+type FilterLeafValues<
+  TLeafCandidate extends string,
+  TNonLeaf extends { [k: string]: TestValue | undefined }
+> = IsNever<TNonLeaf> extends true
+  ? TLeafCandidate
+  : TLeafCandidate extends string
+    ? TLeafCandidate extends keyof TNonLeaf
+      ? never
+      : TLeafCandidate
+    : never;
+
+// this is not 100% accurate since we can't make parallel regions required in the result
+// `TTestValue` doesn't encode this information anyhow for us to be able to do that
+// this is fine for most practical use cases anyway though
+type ToStateValue<TTestValue extends string | TestValue> =
+  | FilterLeafValues<
+      GroupTestValues<TTestValue>['leafCandidates'],
+      GroupTestValues<TTestValue>['nonLeaf']
+    >
+  | (IsNever<GroupTestValues<TTestValue>['nonLeaf']> extends false
+      ? {
+          [K in keyof GroupTestValues<TTestValue>['nonLeaf']]: ToStateValue<
+            NonNullable<GroupTestValues<TTestValue>['nonLeaf'][K]>
+          >;
+        }
+      : never);
 
 export function createMachine<
   TContext extends MachineContext,
@@ -65,6 +111,9 @@ export function createMachine<
   TAction,
   TGuard,
   TDelay,
+  'matchesStates' extends keyof TTypesMeta
+    ? ToStateValue<Cast<TTypesMeta['matchesStates'], TestValue>>
+    : StateValue,
   Prop<
     ResolveTypegenMeta<
       TTypesMeta,
@@ -83,6 +132,7 @@ export function createMachine<
   ResolveTypegenMeta<TTypesMeta, TEvent, TActor, TAction, TGuard, TDelay, TTag>
 > {
   return new StateMachine<
+    any,
     any,
     any,
     any,