+++ packages/xstate-vue/src/useActor.ts
@@ -1,4 +1,5 @@
-import { Ref, shallowRef } from 'vue';
+import isDevelopment from '#is-development';
+import { Ref } from 'vue';
 import {
   ActorOptions,
   ActorRefFrom,
@@ -7,12 +8,12 @@ import {
   Snapshot,
   SnapshotFrom
 } from 'xstate';
-import { UseActorRefRestParams, useActorRef } from './useActorRef.ts';
+import { useActorRef } from './useActorRef.ts';
 import { useSelector } from './useSelector.ts';
 
 export function useActor<TLogic extends AnyActorLogic>(
   actorLogic: TLogic,
-  ...[options = {}]: UseActorRefRestParams<TLogic>
+  options: ActorOptions<TLogic>
 ): {
   snapshot: Ref<SnapshotFrom<TLogic>>;
   send: (event: EventFrom<TLogic>) => void;
@@ -20,22 +21,24 @@ export function useActor<TLogic extends AnyActorLogic>(
 };
 export function useActor(
   actorLogic: AnyActorLogic,
-  options?: ActorOptions<AnyActorLogic>
+  options: ActorOptions<AnyActorLogic> = {}
 ) {
-  if (process.env.NODE_ENV !== 'production') {
-    if ('send' in actorLogic && typeof actorLogic.send === 'function') {
-      throw new Error(
-        `useActor() expects actor logic (e.g. a machine), but received an ActorRef. Use the useSelector(actorRef, ...) hook instead to read the ActorRef's snapshot.`
-      );
-    }
+  if (
+    isDevelopment &&
+    'send' in actorLogic &&
+    typeof actorLogic.send === 'function'
+  ) {
+    throw new Error(
+      `useActor() expects actor logic (e.g. a machine), but received an ActorRef. Use the useSelector(actorRef, ...) hook instead to read the ActorRef's snapshot.`
+    );
   }
 
-  function listener(nextState: Snapshot<unknown>) {
-    snapshot.value = nextState;
+  function listener(nextSnapshot: Snapshot<unknown>) {
+    snapshot.value = nextSnapshot;
   }
 
   const actorRef = useActorRef(actorLogic, options, listener);
-  const snapshot = useSelector(actorRef, (state) => state);
+  const snapshot = useSelector(actorRef, (s) => s);
 
   return {
     snapshot,